feat(03-10): typed-error module + sanctioned logging at the repository error boundary
- Extract WorkspaceErrorCode/WorkspaceError to workspace/errors.ts (imports nothing — db.ts can throw the typed error without a repository→db cycle); repository.ts re-exports both so every existing import path keeps working
- toWorkspaceError now logs every newly-wrapped engine failure through the sanctioned logger (error_code + error_message metadata, fixed compile-time message); already-typed WorkspaceErrors pass through unlogged
- redact allowlist gains error_message with the review-visible rationale: storage-engine failure text only — zod/parse failures keep fixed copy and structurally never reach the wrap boundary (T-03-10-01)
- redact tests cover the new key as a primitive, its default-deny container behavior, and the never-emits-outside-allowlist assertion
